requirements.txt
aiofiles==0.8.0
anyio==3.6.1
asgiref==3.5.1
boto3==1.23.0
botocore==1.26.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Django==4.0.4
django-cors-headers==3.12.0
django-storages==1.12.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
jmespath==1.0.0
numpy==1.22.3
openpyxl==3.0.9
Pillow==9.1.0
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
rfc3986==1.5.0
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
